Diff manifests with raw string
After WPT moves to "Manifest path trie RFC", we don't have the `paths`
field in manifest anymore.

But, we still have file hash in the `items` field; the simplest way to
diff between manifests is to compare the raw string value of the `items`
field. When we have different manifests, we must have different items
field in the manifest json.